Добавил интеграцию с телеграммом

+++ src/model/domain/entities/User.java
@@ -1,14 +1,14 @@
 package model.domain.entities;
 
 public class User {
-    private final Integer userId;
+    private final Long userId;
     private final String platform;
 
-    public User(Integer userId, String platform){
+    public User(Long userId, String platform){
         this.userId = userId;
         this.platform = platform;
     }
 
     public String getPlatform() { return platform; }
-    public Integer getUserId() { return userId; }
+    public Long getUserId() { return userId; }
 }
