+++ src/viewmodel/message/handler/MessageHandler.java
@@ -9,10 +9,10 @@ public interface MessageHandler {
     /**
      * Определяет, может ли этот обработчик взять на себя данное сообщение.
      */
-    boolean canHandle(Integer userId, IRepository<User, Integer> userRepo, String text);
+    boolean canHandle(Long userId, IRepository<User, Long> userRepo, String text);
 
     /**
      * Выполняет логику и возвращает список строк-ответов.
      */
-    List<String> handle(Integer userId, IRepository<User, Integer> userRepo, String text);
+    List<String> handle(Long userId, IRepository<User, Long> userRepo, String text);
 }
