+++ test/test/viewmodel/message/handler/MyVideosHandlerTest.java
@@ -14,9 +14,9 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class MyVideosHandlerTest {
     private IRepository<Video, String> videoRepo;
-    private MyVideosHandler         handler;
-    private static final User TEST_USER = new User(1, "testPlatform");
-    private IRepository<User, Integer>   userRepo;
+    private MyVideosHandler handler;
+    private static final User TEST_USER = new User(1L, "testPlatform");
+    private IRepository<User, Long> userRepo;
 
     @BeforeEach
     void setUp() {
@@ -24,19 +24,18 @@ class MyVideosHandlerTest {
         userRepo  = new InMemoryRepository<>(User::getUserId);
         userRepo.save(TEST_USER);
         handler   = new MyVideosHandler(videoRepo);
-
     }
 
     @Test
     void canHandle_onlyListVideosCommand() {
-        assertTrue(handler.canHandle(1, userRepo, "/MyVideos"));
-        assertTrue(handler.canHandle(1, userRepo, "/MyVideos     "));
-        assertFalse(handler.canHandle(1, userRepo, "/lists"));
+        assertTrue(handler.canHandle(1L, userRepo, "/MyVideos"));
+        assertTrue(handler.canHandle(1L, userRepo, "/MyVideos     "));
+        assertFalse(handler.canHandle(1L, userRepo, "/lists"));
     }
 
     @Test
     void handle_noVideos_returnsEmptyNotice() {
-        List<String> resp = handler.handle(42, userRepo, "/MyVideos");
+        List<String> resp = handler.handle(42L, userRepo, "/MyVideos");
         assertEquals(1, resp.size());
         assertEquals("У вас ещё нет добавленных видео.", resp.get(0));
     }
@@ -44,26 +43,27 @@ class MyVideosHandlerTest {
     @Test
     void handle_withVideos_filtersByUser() {
         Video v1 = new Video("u1","p","id1",0,"t");
-        v1.setUserAdded(5);
+        v1.setUserAdded(5L);
         Video v2 = new Video("u2","p","id2",0,"t");
-        v2.setUserAdded(6);
+        v2.setUserAdded(6L);
         videoRepo.save(v1);
         videoRepo.save(v2);
 
-        List<String> resp = handler.handle(5, userRepo, "/MyVideos");
+        List<String> resp = handler.handle(5L, userRepo, "/MyVideos");
         assertEquals(1, resp.size());
         assertTrue(resp.get(0).contains("u1"));
     }
+
     @Test
     void handle_multipleVideosForSameUser_listsAll() {
         Video v1 = new Video("u1","p","id1",0,"t");
-        v1.setUserAdded(42);
+        v1.setUserAdded(42L);
         Video v2 = new Video("u2","p","id2",0,"t");
-        v2.setUserAdded(42);
+        v2.setUserAdded(42L);
         videoRepo.save(v1);
         videoRepo.save(v2);
 
-        List<String> resp = handler.handle(42, userRepo, "/MyVideos");
+        List<String> resp = handler.handle(42L, userRepo, "/MyVideos");
         assertEquals(2, resp.size(),
                 "Оба видео одного пользователя должны быть выведены");
         assertTrue(resp.stream().anyMatch(s -> s.contains("u1")));
@@ -73,10 +73,10 @@ class MyVideosHandlerTest {
     @Test
     void handle_userWithNoVideos_returnsEmptyNotice() {
         Video v1 = new Video("u1","p","id1",0,"t");
-        v1.setUserAdded(99);
+        v1.setUserAdded(99L);
         videoRepo.save(v1);
 
-        List<String> resp = handler.handle(42, userRepo, "/MyVideos");
+        List<String> resp = handler.handle(42L, userRepo, "/MyVideos");
         assertEquals(1, resp.size(),
                 "Если у пользователя нет видео, должен быть один ответ");
         assertEquals("У вас ещё нет добавленных видео.", resp.get(0));
@@ -84,20 +84,19 @@ class MyVideosHandlerTest {
 
     @Test
     void handle_withNullMessage_returnsEmptyNotice() {
-        List<String> resp = handler.handle(42, userRepo, null);
+        List<String> resp = handler.handle(42L, userRepo, null);
         assertEquals(1, resp.size(),
                 "При null-сообщении должен вернуться один ответ");
         assertEquals("У вас ещё нет добавленных видео.", resp.get(0));
     }
 
-
     @Test
     void handle_otherUsersVideos_notIncluded() {
         Video v1 = new Video("u1","p","id1",0,"t");
-        v1.setUserAdded(100);
+        v1.setUserAdded(100L);
         videoRepo.save(v1);
 
-        List<String> resp = handler.handle(200, userRepo, "/MyVideos");
+        List<String> resp = handler.handle(200L, userRepo, "/MyVideos");
         assertEquals(1, resp.size(),
                 "Если у пользователя нет своих видео, должен быть один ответ");
         assertEquals("У вас ещё нет добавленных видео.", resp.get(0));
@@ -105,10 +104,9 @@ class MyVideosHandlerTest {
 
     @Test
     void handle_withWhitespaceMessage_returnsEmptyNotice() {
-        List<String> resp = handler.handle(42, userRepo, "   ");
+        List<String> resp = handler.handle(42L, userRepo, "   ");
         assertEquals(1, resp.size(),
                 "При сообщении из пробелов должен вернуться один ответ");
         assertEquals("У вас ещё нет добавленных видео.", resp.get(0));
     }
-
 }
